Lower version requirement on FreeBSD
FreeBSD is not a supported platform of libusb. The native USB access
library on FreeBSD is libusb20, which provides both the non-portable
FreeBSD-specific API and wrapper functions that are mostly compatible
with libusb 1.0 API. It is an integral part of FreeBSD and the version
included in the oldest supported FreeBSD release reports 1.0.9 in its
libusb-1.0.pc.