Simplify platform ID building and provide fallback for libusb_get_parent
FreeBSD doesn't have libusb_get_parent. Fortunately, we can find the
parent device manually using the bus number and port numbers. However,
calling libusb_get_port_numbers on FreeBSD requires calling libusb_open
because it needs an open file descriptor to do ioctls.

To avoid making system calls every time we need to know the parent
device, g_usb_device_initable_init calls libusb_get_port_numbers and
store the list of port numbers in GUsbDevice itself. When we need the
parent device, we can call g_usb_context_get_devices to get the list of
all devices and check the bus number and port numbers. It is still
slower than storing a parent pointer in the object itself, which is what
libusb on Linux currently does, but I think the number of USB devices is
usually small so the performance impact can be ignored.

This change also simplifies the process of platform ID building. Since
getting the parent requires the list of all devices on FreeBSD, we are
no longer able to call it during the initialization of GUsbDevice.
Therefore, we simply construct the platform ID from tne bus number and
port numbers, which is the same as what the code intended to do.

Note that this commit does change the value of platform ID. It is
actually a bug fix because the condition 'str->len == 0' in
g_usb_device_build_platform_id_cb could not be true and the port number,
which is always zero because it is unavailable on root hubs, is used
instead of the bus number. This means the platform ID was always started
with 'usb:00' and it was not unique on the same host.