Root hubs on FreeBSD have vid and pid set to zero
This doesn't really make the test work on FreeBSD. You still have to
comment out g_object_unref calls or it will hang in g_usb_context_dispose.

The first problem is that libusb_handle_events never returns. It polls
on an internal pipe called 'ctrl_pipe' with default timeout -1. It can
only be unblocked by opening or closing a device, or doing some data
transfer works. Even if we replace libusb_handle_events with the new
libusb_handle_events_timeout_completed function, it can still hang for
several seconds, depending on the timeout we specify.

The second problem is that libusb_exit can block for up to 4 seconds.
When libusb_hotplug_register_callback is called, it starts a thread
which sleep 4 seconds before doing any work. libusb_exit waits for the
thread and there is no way to interrupt the sleep.